Rename positioned graph module and refactor tests
